Flattening overrides and defaults into config.
With deep access you'd end up in odd situation where setting an override
would appear to work because it'd be shadowed by an ancestor in the configuration
array.

Maintaining individual configuration stores here offers a degree of insight
but I feel the most important piece there is simply being able to log when
/ where different config keys are set. That functionality already exists.

Instead, I've decided to maintain a single point of truth.